Stage 6.
Designed Stage 6 with a few puzzles and 2 hidden paths.  The hidden paths are easier than following the normal path.  You are able to skip 2 puzzles if you can find the hidden path.  I did not test all puzzles, yet.  Did not make the tunnel to stage 6.

Need tunnel and camera to Stage 6.